### 2018-12-09 #271 @melvidoni
### Editor checks:

---

#### Editor comments

Thank you for the submission @cboettig! Here is the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). There is no need to address them now, as this is information for the reviewers.

```
── GP virtuoso ──────────────────────────────────────────────────────────────────

It is good practice to

    R/odbc.R:55:NA
    R/odbc.R:105:NA
    R/odbc.R:106:NA
    R/odbc.R:107:NA
    R/odbc.R:108:NA
    ... and 95 more lines

  ✖ add a "URL" field to DESCRIPTION. It helps users find
    information about your package online. If your package does not
    have a homepage, add an URL to GitHub, or the CRAN package package
    page.
  ✖ add a "BugReports" field to DESCRIPTION, and point it to
    a bug tracker. Many online code hosting services provide bug
    trackers for free, https://github.com, https://gitlab.com, etc.

    inst\examples\docker.R:13:1
    inst\examples\json.R:9:1
    R\virtuoso.R:32:1
    R\virtuoso.R:42:1
    R\virtuoso.R:44:1
    ... and 6 more lines

  ✖ fix this R CMD check ERROR: Packages suggested but not
    available: 'nycflights13' 'jsonld' 'rdftools' 'spelling' The
    suggested packages are required for a complete check. Checking can
    be attempted without them by setting the environment variable
    _R_CHECK_FORCE_SUGGESTS_ to a false value. See section 'The
    DESCRIPTION file' in the 'Writing R Extensions' manual.
─────────────────────────────────────────────────────────────────────────────────
```

I'll be seeking for reviewers now, and update the information.

---

Reviewers:
Due date:
