### 2018-12-10 110b1eb
A few tweaks for goodpractice
