+++ DESCRIPTION
@@ -27,6 +27,8 @@ Description: Virtuoso is a high-performance "universal server," which can act
              hundreds of times faster than the comparable imports using existing R tools,
              including rdflib and redland packages.  
 License: MIT + file LICENSE
+URL: https://github.com/cboettig/virtuoso
+BugReports: https://github.com/cboettig/virtuoso/issues
 Encoding: UTF-8
 LazyData: true
 Imports: 
