+++ NAMESPACE
@@ -27,5 +27,4 @@ importFrom(odbc,odbc)
 importFrom(processx,process)
 importFrom(processx,run)
 importFrom(rappdirs,app_dir)
-importFrom(utils,askYesNo)
 importFrom(utils,read.table)
