### 2018-12-11 0886bd9
A few tweaks for goodpractice (#21)
As suggested by ropensci onboarding editor.
