### 2018-12-23 0f86d77
more goodpractice fixes (mostly whitespace)
