### 2018-12-23 d1aaf65
add hex, update pkgdown
