+++ README.Rmd
@@ -2,6 +2,9 @@
 output: github_document
 ---
 
+
+# virtuoso <img src="man/figures/logo.svg" align="right" alt="" width="120" />
+
 [![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
 [![Travis build status](https://travis-ci.org/cboettig/virtuoso.svg?branch=master)](https://travis-ci.org/cboettig/virtuoso)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/virtuoso?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/virtuoso)
@@ -18,7 +21,7 @@ knitr::opts_chunk$set(
   out.width = "100%"
 )
 ```
-# virtuoso
+
 
 The goal of virtuoso is to provide an easy interface to Virtuoso RDF database from R. 
 
