### 2018-12-27 5cdba8b
Just a hex and updated pkgdown (#22)
* A few tweaks for goodpractice

* more goodpractice fixes (mostly whitespace)

* add hex, update pkgdown

* has favicons

* update buildignore

* as regex
