+++ tests/testthat/test-vos_install-errors.R
@@ -36,15 +36,15 @@ test_that("We get errors running windows installer on non-windows",
 
 
 test_that("We get errors running mac installer on non-mac",
-          {
-            skip_on_cran()
-            skip_on_os("mac")
-            skip_on_os("windows")
-#            expect_error(virtuoso:::vos_install_osx(prompt = FALSE))
-#            expect_error(virtuoso:::vos_install_osx(brew = TRUE, prompt = FALSE))
+  {
+  skip_on_cran()
+  skip_on_os("mac")
+  skip_on_os("windows")
+#   expect_error(virtuoso:::vos_install_osx(prompt = FALSE))
+#   expect_error(virtuoso:::vos_install_osx(brew = TRUE, prompt = FALSE))
 
-            ## remarkably / scarily, this doesn't throw an error...
-            #expect_error(virtuoso:::install_brew())
+  ## remarkably / scarily, this doesn't throw an error...
+  #expect_error(virtuoso:::install_brew())
 
 
           })
