### 2018-12-27 da8863a
drop rdftools dependency
