### 2019-01-06 #271 @czeildi
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

In general the package has extensive documentation and easy-to-follow code, it is a pleasure to review!
I included checkboxes for the comments which I feel are relatively important. The other comments are rather for consideration and would not block my recommendation for approval.

I tried the package on my mac.

- [ ] Add examples to function help pages. They should be probably in `\dontrun{}` as they have heavy dependencies. Even if the example is just restating how the function can be called I think they are useful, many users first scroll to the examples section in the documentation.
- [ ] Add package documentation, skeleton could be added by `usethis::use_package_doc()`.
- [ ] [`vos_count_triples()`](https://cboettig.github.io/virtuoso/reference/vos_count_triples.html) should be removed from pkgdown reference until it is implemented
- [ ] fix `jsonlite::read_json("https://raw.githubusercontent.com/ropensci/roregistry/ex/codemeta.json")` in datalake article, related comment which I found is [here](https://github.com/ropensci/roregistry/issues/23#issuecomment-436271532).
- For me the following line did not run in the datalake vignette: `system.time(
  vos_import(con, c("flights.nq", "planes.nq", "airlines.nq"))
)`. I have not yet investigated further, it is possibly a problem specific to my machine, but please confirm that it is working for you with a clean install.
- For first time users it can be helpful to group functions for the pkgdown reference site. I could imagine 3 groups: getting started, working with data and advanced usage / troubleshooting. For guidance see the [Grouping functions in the reference section](https://ropensci.github.io/dev_guide/building.html#website) in the ropensci guide.
- [ ] Links do not work in function help pages, this is fixed by adding the following line to the `DESCRIPTION` file: `Roxygen: list(markdown = TRUE)`
- `TRUE`, `FALSE` should be marked as code in `vos_destroy_all()` documentation, in `vos_log()` in `just_errors` `true` should be capital and shown as code.
- In details section of `vos_import()` after list of all extensions there should be a newline.
- Remove one \` from double \` in `vos_import.R` and `vos_status.R`
- `vos_odbcinst()`: ODBC Uses a: uses should be lower case
- In `DESCRIPTION` file: The virtuoso package provides or the virtuoso R package provides
- spell check errors in `README.Rmd` line 80: Remember, because, convenient
- [ ] Add a getting started vignette (called virtuoso so that it shows up as Getting Started in pkgdown website): this could have the same content as the second part of the `README`.
- In datalake article: Change sentence to this: 'First, we must represent any primary or foreign keys in any table as URIs and not literal integers or strings:'
- [ ] In `vos_install()` if I see correctly, this is a general function and the documentation title is incorrect that it is only for mac. This is in conflict with readme on what platforms are fully supported.

#### Usage

- In `vos_install()` if Virtuoso is already installed, for me the message could be slightly cleaner with sg like this: Virtuoso installation already present so no installation happens.
- [ ] I think `vos_kill` should not throw error if there is nothing to kill.
- In `vos_status()` I would not give a warning, only a message or plain return value if no process. A warning indicates sg that should be fixed, I might just want to know whether I should start or kill virtuoso.
- In my experience `NULL` is more widespread and thus easier to understand as default parameter value than `NA` in `vos_kill()`. Reference [here](https://stackoverflow.com/questions/29619431/should-i-use-null-or-na-as-default-values-for-optional-function-parameters).
- [ ] `vos_delete_db()` fails with a non-informative error message in non-interactive environment or if `ask = FALSE`.

- [ ] For installation and other code, hard coding the version of Virtuoso open source does not seem future proof. Consider either installing always the latest version or making the version a parameter or environment variable. Or at least make this clear in README that the code works with a specific version if local server is considered.
- react to your own FIXME comments in code if relevant ( :) )
- linux branch in `vos_uninstall()`: consider converting to `message` (consistent with `vos_uninstall_osx()`) or error (consistent with `vos_install_linux()`)
- I see that in a recent commit you removed `rdftools` from Suggests. For someone rebuilding articles locally it would be nice to add sg like this to the article: `if (!requre('rdftools')) stop('please install...')`. Although this might only be relevant for this review process: until I found your issue already discussing this my first thought was that this package should be added to `Suggests`.
- I am not sure about this but would it make sense to include addons from travis as `SystemRequirements` in `DESCRIPTION`?
- Use `testthat::setup` and `teardown` for `vos_start` and `vos_kill` in `test-vos_query` as it is more transparent and it will ensure `vos_kill` is run even if there is some error during running the tests.
```r
testthat::teardown(vos_kill())
```
- sleep should only happen in test setup if installation actually happens, otherwise we wait 20 seconds for nothing
- consider specifying message (or message pattern) in `expect_warning` in order to avoid there being a completely different warning
- consider mocking as a way to add tests to functions which are difficult to test (such as platform dependent functions)
- use codemetar as it is recommended by ropensci
- ask and prompt function argument name is used for the same purposes, consider unifying these
- `is_interactive()` function is defined, but not used at all places instead of `interactive()`
- fs is already imported -> might consider https://fs.r-lib.org/reference/path_package.html instead of system.file as it cannot fail silently

##### style

In general the code is very easily readable. However, code style is not fully consistent. Examples: space after `if`, space before `{`, indentation of multiple function arguments, using simple or double quotes. Making these consistent can help future potential contributors as it is clear what style they should follow. (I am not sure whether it is possible to configure `styler` to automatically ensure a style close to your current style.)

- `virtuoso:::` in tests is a bit confusing to me and I believe it is not necessary
- In `vos_status()` details: \`\[vos_log()\]\`\` -> \[\`vos_log()\`\] for consistency
