### 2019-01-21 #271 @edgararuiz
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

#### Test results

4 tests fail. I tested it on my Windows laptop.  Something I noticed is that I
have to do `vos_install()` and then `vos_start()` every time I tried to use
`virtuoso` in a new R session. I believe that explains the issues brougth up by
`testthat`
  ```
  Loading virtuoso
  Testing virtuoso
  Running one-time install of Virtuoso for tests...
  v | OK F W S | Context
  v |  5       | utilities
  v |  2       | vos_configure
  x |  2 1   2 | vos install errors [0.4 s]
  ------------------------------------------------------
  test-vos_install-errors.R:12: error: We can download the windows installer
  HTTP error 404.
  1: virtuoso:::download_windows_installer() at C:\Users\edgar\Documents\virtuoso/tests/testthat/test-vos_install-errors.R:12
  2: curl::curl_download(download_url, installer) at C:/Users/edgar/Documents/virtuoso/R/vos_install_windows.R:12

  test-vos_install-errors.R:32: skip: We get errors running windows installer on non-windows
  On windows

  test-vos_install-errors.R:42: skip: We get errors running mac installer on non-mac
  On windows
  ------------------------------------------------------
  v |  1     1 | vos install
  ------------------------------------------------------
  test-vos_install.R:13: skip: We can download installers
  On windows
  ------------------------------------------------------
  v | 11       | Detecting ODBC Drivers
  x |  0 1     | test vos_query [33.2 s]
  ------------------------------------------------------
  test-vos_query.R:3: error: (unknown)
  No such process, pid 6812, ???
  1: vos_start() at C:\Users\edgar\Documents\virtuoso/tests/testthat/test-vos_query.R:3
  2: vos_status(p, wait = wait) at C:/Users/edgar/Documents/virtuoso/R/vos_start.R:44
  3: p$get_status() at C:/Users/edgar/Documents/virtuoso/R/vos_status.R:36
  4: ps_method(ps::ps_status, self)
  5: fun(ps::ps_handle(self$get_pid(), self$get_start_time()))
  ------------------------------------------------------
  x |  2 1     | vos_start
  ------------------------------------------------------
  test-vos_start.R:13: error: we can start a vos server and check status
  No such process, pid 6812, ???
  1: expect_true(p$get_status() %in% c("sleeping", "running")) at C:\Users\edgar\Documents\virtuoso/tests/testthat/test-vos_start.R:13
  2: quasi_label(enquo(object), label)
  3: eval_bare(get_expr(quo), get_env(quo))
  4: p$get_status() %in% c("sleeping", "running")
  5: p$get_status()
  6: ps_method(ps::ps_status, self)
  7: fun(ps::ps_handle(self$get_pid(), self$get_start_time()))
  ------------------------------------------------------
  x |  1 1     | tests that do not need a server connection
  ------------------------------------------------------
  test-without-vos.R:7: failure: vos_process errors when not cached
  `vos_process()` did not produce any warnings.
  ------------------------------------------------------

  == Results ===========================================

  OK:       24
  ```

Some of the funcitonality works, but not all of what is shown in the README.  Here is a `reprex` of what I see:

``` r
library(virtuoso)
vos_start()
#> Error in vos_start(): Virtuoso installation not detected.  See vos_install()
vos_install()
vos_start()
#> PROCESS 'virtuoso-t', running, pid 2884.
#> Server is now starting up, this may take a few seconds...
#> latest log entry: 11:06:25 Server exiting
#> Error: No such process, pid 2884, ???
con <- vos_connect()
DBI::dbGetQuery(con, "SPARQL SELECT * WHERE { ?s ?p ?o } LIMIT 4")
#>                                                                              s
#> 1                   http://www.openlinksw.com/virtrdf-data-formats#default-iid
#> 2          http://www.openlinksw.com/virtrdf-data-formats#default-iid-nullable
#> 3          http://www.openlinksw.com/virtrdf-data-formats#default-iid-nonblank
#> 4 http://www.openlinksw.com/virtrdf-data-formats#default-iid-nonblank-nullable
#>                                                 p
#> 1 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 2 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 3 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#> 4 http://www.w3.org/1999/02/22-rdf-syntax-ns#type
#>                                                         o
#> 1 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 2 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 3 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
#> 4 http://www.openlinksw.com/schemas/virtrdf#QuadMapFormat
example <- system.file("extdata", "person.nq", package = "virtuoso")
vos_import(con, example)
#> Error: <SQL> 'ld_dir('C:/Users/edgar/AppData/Local/Virtuoso/Virtuoso/Cache/c990c6c99cd92278bde129f6fedd5b65', '*', 'rdflib')'
#>   nanodbc/nanodbc.cpp:1587: 42000: [OpenLink][Virtuoso ODBC Driver][Virtuoso Server]FA003: Access to 'C:\Users\edgar\AppData\Local\Virtuoso\Virtuoso\Cache\c990c6c99cd92278bde129f6fedd5b65' is denied due to access control in ini file
vos_query(con,
          "SELECT ?p ?o
 WHERE { ?s ?p ?o .
        ?s a <http://schema.org/Person>
       }")
#> [1] p o
#> <0 rows> (or 0-length row.names)
```
