### 2019-01-22 #271 @melvidoni
@cboettig Reviews have been completed. Please, perform the corresponding changes, and comment again answering the reviewers with your updates.
