### 2019-01-22 #271 @cboettig
Thanks both, this is super helpful.

@edgararuiz , I'm traveling at the the moment so won't have access to my Windows machine until next week when I can poke around more.  Definitely should not have to re-install each time though! A few questions meanwhile that can help me debug:

- When you run `vos_install()`, does this open the standard Windows installation dialog that asks you to install Virtuoso?  Can you try again with `vos_install(prompt=FALSE)`?  This will run an non-interactive version of the installer.  You should not have to reinstall unless you run `vos_uninstall()` or manually uninstall Virtuoso.

- After you run `vos_start()`, can you let me know what it says on the logfile:

```r
readLines( file.path(virtuoso:::vos_logdir(), "virtuoso.log"))
```

From the log above, it looks like Virtuoso is installed and starts up but then shuts down.  I think this is due to a file permission error somewhere, but I'm not quite sure where.

- Can you tell me what you get from:

```r
virtuoso:::find_virtuoso_ini()
virtuoso:::vos_db()
virtuoso:::vos_logdir()
```

Each of those should return file path -- can you determine if your user has write permissions to each of those locations?

Thanks!
