### 2019-01-29 f7ed350
edits for rOpenSci review
