+++ R/virtuoso.R
@@ -1,27 +1,15 @@
-## Some possible helper routines for common requests.
+## Some additional possible helper routines for common requests are shown here.
+## Currently not fully developed or tested, and thus not exported.
 
 #' Clear all triples from a graph
 #'
 #' @details NOTE: after clearing a graph, re-running the bulk
-#' importer may not re-import triples.
+#' importer may refuse to re-import triples.
 #' @inheritParams vos_import
 vos_clear_graph <- function(con, graph = "rdflib"){
   DBI::dbGetQuery(con, paste0("SPARQL CLEAR GRAPH <", graph, ">"))
 }
 
-#' Delete Virtuoso Database
-#'
-#' delete the entire Virtuoso database for a fresh start.
-#' @param ask prompt before deleting?
-#' @param db_dir location of the directory to delete
-#' @export
-vos_delete_db <- function(ask = interactive(),
-                          db_dir = vos_db()){
-  if(ask)
-    continue <- askYesNo("Are you sure?")
-  if(continue)
-    unlink(db_dir, recursive = TRUE)
-}
 
 
 #' List graphs
@@ -45,6 +33,7 @@ vos_list_graphs <- function(con){
 #' count triples
 #'
 #' @inheritParams vos_import
+#' @noRd
 vos_count_triples <- function(con, graph = NULL){
 
   ## Official query method below.  Not sure why these return
