+++ R/vos_import.R
@@ -1,14 +1,18 @@
 
 
-#' Bulk import of nquads
+#' Bulk Import of RDF triples
 #'
-#' @param con a ODBC connection to Virtuoso, from [`vos_connect()`]
+#' While triples data can be added one by one over SPARQL queries,
+#' Virtuoso bulk import is by far the fastest way to import large
+#' triplestores in the database.
+#'
+#' @param con a ODBC connection to Virtuoso, from [vos_connect()]
 #' @param files paths to files to be imported
 #' @param wd Alternatively, can specify directory and globbing pattern
 #'  to import. Note that in this case, wd must be in (or a subdir of)
 #'  the `AllowedDirs` list of `virtuoso.ini` file created by
-#'  [`vos_configure()`]. By default, this includes the working directory
-#'  where you called [`vos_start()`] or [`vos_configure()`].
+#'  [vos_configure()]. By default, this includes the working directory
+#'  where you called [vos_start()] or [vos_configure()].
 #' @param glob A wildcard aka globbing pattern (e.g. `"*.nq"``).
 #' @param graph Name (technically URI) for a graph in the database.
 #'  Can leave as default. If a graph is already specified by the
@@ -38,11 +42,21 @@
 #'  - `.trig`
 #'  - `.ttl`
 #'  - `.xml`
+#'
 #'  Any of these can optionally be gzipped (with a `.gz` extension).
 #' @references <http://vos.openlinksw.com/owiki/wiki/VOS/VirtBulkRDFLoader>
 #' @importFrom digest digest
 #' @importFrom fs path_abs
 #' @export
+#' @examples \dontrun{
+#' vos_start()
+#' con <- vos_connect()
+#'
+#' example <- system.file("extdata", "person.nq", package = "virtuoso")
+#' vos_import(con, example)
+#'
+#'
+#' }
 vos_import <- function(con,
                        files = NULL,
                        wd = ".",
