### 2019-01-30 e7bc947
go directly to nquads
