+++ vignettes/articles/datalake.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 
 ```{r}
 library(virtuoso)
-library(rdftools) # for write_nquads
+library(rdftools) # for write_nquads.  Needs: install_github("cboettig/rdftools")
 library(dplyr)
 
 ```
@@ -35,13 +35,15 @@ library(nycflights13)
 We start up our Virtuoso server, wait for it to come up, and then connect:
 
 ```{r install}
-myserver <- vos_start()
-Sys.sleep(5)
+vos_start()
+```
+
+```{r}
 con <- vos_connect()
 ```
 
 
-We can represent any data as RDF with a little care.  For instance, consider the `nycflights13` data. First, we must represent any primary or foreign keys in any table must be as URIs and not literal integers or strings: 
+We can represent any data as RDF with a little care.  For instance, consider the `nycflights13` data. First, we must represent any primary or foreign keys in any table as URIs, indicated by a prefix, and not by bare strings:
 
 ```{r as_uri}
 uri_flights <- flights %>% 
@@ -49,7 +51,7 @@ uri_flights <- flights %>%
          carrier = paste0("airlines:", carrier))
 ```
 
-We write the `data.frame`s out as nquads.  Recall that each cell of a `data.frame` can be represented as a triple, in which the column is the predicate, the primary key (or row number) the subject, and the cell value the object.  We turn column names and primary keys into URIs using a prefix based on the table name.  (Note that `rdflib` does this conversion by merely munging cells and calling `write.table`, it is not a standard `redland` library transform).
+We write the `data.frame`s out as nquads.  Recall that each cell of a `data.frame` can be represented as a triple, in which the column is the predicate, the primary key (or row number) the subject, and the cell value the object.  We turn column names and primary keys into URIs using a prefix based on the table name. 
 
 ```{r}
 write_nquads(airlines,  "airlines.nq", key = "carrier", prefix = "airlines:")
@@ -99,15 +101,19 @@ Transform JSON (or list data) into triples.  In this case, we have a large JSON 
 containing metadata on all rOpenSci packages:
 
 ```{r}
-x <- jsonlite::read_json("https://raw.githubusercontent.com/ropensci/roregistry/ex/codemeta.json")
-write_nquads(x, "ropensci.nq", prefix = "http://schema.org/")
+library(jsonlite)
+library(jsonld)
+
+download.file("https://raw.githubusercontent.com/ropensci/roregistry/gh-pages/raw_cm.json", "raw_cm.json")
+nq <- jsonld::jsonld_to_rdf("raw_cm.json") 
+writeLines(nq, "ro.nq.gz")
 ```
 
 
 And bulk-import
 
 ```{r}
-vos_import(con, "ropensci.nq")
+vos_import(con, "ro.nq.gz")
 ```
 
 Find all packages where "Carl Boettiger" is an "author", and return:
@@ -128,36 +134,17 @@ SELECT DISTINCT ?package ?license ?coauthor
  ?coauth schema:familyName ?coauthor
 }"
 
-vos_query(con, query) %>% 
-as_tibble() %>% 
-mutate(license = basename(license)) # Tidy up URIs into names
-```
-
-
-```{r}
-query <-
-  rdftools:::sparql_op() %>%
-  rdftools:::select("identifier", "license", prefix = "http://schema.org/") %>%
-  rdftools:::filter(author.familyName == "Boettiger",
-         author.givenName == "Carl",
-        prefix = "http://schema.org/") %>%
-  rdftools:::sparql_build()
-
-vos_query(con, query)
-
+vos_query(con, query) %>% distinct() %>%
[…]
