### 2019-01-31 bfbbf5e
update docs
- add rOpenSci badge
- tweak datalake example
