### 2019-01-31 #271 @edgararuiz
Hi, sorry for the delay in the response, I've been out of town

In a new R session, when I run `vos_install()` it just runs silently and quickly.  And then when I run `vos_start()` it seems to start ok:
```r
readLines( file.path(virtuoso:::vos_logdir(), "virtuoso.log"))
 [1] ""
 [2] "\t\tThu Jan 31 2019"
 [3] "15:45:58 [Using virtuoso.ini in C:\\Users\\edgar\\AppData\\Local\\Virtuoso\\Virtuoso]"
 [4] "15:45:58 { Loading plugin 1: Type `plain', file `wikiv' in `C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting'"
 [5] "15:45:58   WikiV version 0.6 from OpenLink Software"
 [6] "15:45:58   Support functions for WikiV collaboration tool"
 [7] "15:45:58   SUCCESS plugin 1: loaded from C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting\\wikiv.dll }"
 [8] "15:45:58 { Loading plugin 2: Type `plain', file `mediawiki' in `C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting'"
 [9] "15:45:58   MediaWiki version 0.1 from OpenLink Software"
[10] "15:45:58   Support functions for MediaWiki collaboration tool"
[11] "15:45:58   SUCCESS plugin 2: loaded from C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting\\mediawiki.dll }"
[12] "15:45:58 { Loading plugin 3: Type `plain', file `creolewiki' in `C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting'"
[13] "15:45:58   CreoleWiki version 0.1 from OpenLink Software"
[14] "15:45:58   Support functions for CreoleWiki collaboration tool"
[15] "15:45:58   SUCCESS plugin 3: loaded from C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting\\creolewiki.dll }"
[16] "15:45:58 { Loading plugin 4: Type `plain', file `im' in `C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting'"
[17] "15:45:58   IM version 0.6 from OpenLink Software"
[18] "15:45:58   Support functions for Image Magick 6.9.9"
[19] "15:45:58   SUCCESS plugin 4: loaded from C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting\\im.dll }"
[20] "15:45:58 { Loading plugin 5: Type `plain', file `wbxml2' in `C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting'"
[21] "15:45:58   WBXML2 version 0.9 from OpenLink Software"
[22] "15:45:58   Support functions for WBXML2 0.9.2 Library"
[23] "15:45:58   SUCCESS plugin 5: loaded from C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\hosting\\wbxml2.dll }"
[24] "15:45:58 OpenLink Virtuoso Universal Server"
[25] "15:45:58 Version 07.20.3217-threads for Win64 as of Sep  6 2017"
[26] "15:45:58 uses parts of OpenSSL, PCRE, Html Tidy"
[27] "15:45:58 Database version 3126"
[28] "15:45:58 SQL Optimizer enabled (max 1000 layouts)"
[29] "15:45:59 Compiler unit is timed at 0.000137 msec"
[30] "15:46:01 Roll forward started"
[31] "15:46:01     3 transactions, 185 bytes replayed (100 %)"
[32] "15:46:01 Roll forward complete"
[33] "15:46:02 PL LOG: Can't get list of vad packages in C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\vad/"
[34] "15:46:02 Checkpoint started"
[35] "15:46:02 Checkpoint finished, log reused"
[36] "15:46:02 HTTP/WebDAV server online at 8890"
[37] "15:46:02 Server online at 1111 (pid 8120)"
```

If I restart R, and just run `vos_start()` I get:

```r
Error in vos_start() :
  Virtuoso installation not detected.  See vos_install()
```

The log does not have any new entries.

Here are the results of the other commands you requested to run:

```r
> virtuoso:::find_virtuoso_ini()
[1] "C:\\Program Files\\OpenLink Software\\Virtuoso OpenSource 7.20\\database\\virtuoso.ini"
> virtuoso:::vos_db()
[1] "C:\\Users\\edgar\\AppData\\Local\\Virtuoso\\Virtuoso"
> virtuoso:::vos_logdir()
[1] "C:\\Users\\edgar\\AppData\\Local\\Virtuoso\\Virtuoso\\Logs"
```
