### 2019-01-31 #271 @cboettig
@edgararuiz No worries and thanks!  I think I have resolved some of these issues now, though I still have a few edge cases to work out, (including appveyor).  Can you test the `ropensci-review` branch, e.g. with

```r
install_github("cboettig/virtuoso@ropensci-review")
```

There's still one case I'm working on, if virtuoso server is already running but not controlled by the R session (e.g. because the user doesn't uncheck the box or an R session doesn't shut down properly after starting one), then currently `vos_start()` will fail to start.  I believe I can now solve this, so that R will simply grab the active virtuoso process, but needs a bit more testing.

Thanks for this report.  I'll ping soon when I have a few more things ironed out.
