### 2019-02-01 62927d3
WIP migrate to ps-based service handling
Using ps package directly instead of caching the processx handle allows us to discover and connect to any running virtuoso process (modulo user permissions).  This streamlines handling of the server process, including stopping and starting a server process started outside of R.
