+++ R/vos_status.R
@@ -11,6 +11,7 @@
 #'  - "running" Server is up and accepting queries.
 #'  - "sleeping" Server is up and accepting queries.
 #'
+#' @importFrom ps ps_status
 #' @export
 #' @examples \dontrun{
 #' vos_status()
@@ -18,18 +19,13 @@
 vos_status <- function(p = NA, wait = 10){
 
   p <- vos_process(p)
-  if(!inherits(p, "process")) return("not detected")
+  if(!inherits(p, "ps_handle")) return(invisible(NULL))
 
-  if (!p$is_alive()) {
-    message(paste("Server is not alive, please restart. Server log: \n\n",
-                  vos_log(p, collapse = "\n")), call. = FALSE)
-    return("dead")
-  }
+  status <- ps::ps_status(p)
 
-  if (!(p$get_status() %in% c("running", "sleeping"))) # stopped,
-    return(p$get_status())
+  if ( !(status %in% c("running", "sleeping")) )
+    return(status)
 
-  Sys.sleep(1)
   log <- vos_log(p, collapse = "\n")
   tries <- 0
   up <- grepl("Server online at", log)
@@ -43,5 +39,5 @@ vos_status <- function(p = NA, wait = 10){
   log <- vos_log(p)
   message(paste("latest log entry:", log[length(log)]))
 
-  p$get_status()
+  status
 }
