+++ R/vos_start.R
@@ -51,10 +51,11 @@ vos_start <- function(ini = NULL, wait = 30){
 
 
   ## Check for cached process
-  p <- mget("virtuoso_process", envir = virtuoso_cache, ifnotfound = NA)[[1]]
-  if (inherits(p, "process")) {
-    message(paste("Found existing process:", p$format()))
-    return(p)
+  p <- vos_process()
+
+  if (inherits(p, "ps_handle")) {
+    message(paste("Returning existing virtuoso-t process with pid:", ps_pid(p)))
+    return(invisible(p))
   }
 
   ## Prepare a virtuoso.ini configuration file if one is not provided.
@@ -68,8 +69,6 @@ vos_start <- function(ini = NULL, wait = 30){
                              stderr = err, stdout = "|",
                              cleanup = TRUE)
 
-  ## Cache the process so we can control it later.
-  assign("virtuoso_process", p, envir = virtuoso_cache)
 
   ## Wait for status
   message(p$format())
