+++ NAMESPACE
@@ -26,5 +26,9 @@ importFrom(ini,write.ini)
 importFrom(odbc,odbc)
 importFrom(processx,process)
 importFrom(processx,run)
+importFrom(ps,ps)
+importFrom(ps,ps_handle)
+importFrom(ps,ps_kill)
+importFrom(ps,ps_status)
 importFrom(rappdirs,app_dir)
 importFrom(utils,read.table)
