+++ DESCRIPTION
@@ -40,7 +40,8 @@ Imports:
     rappdirs,
     curl,
     fs,
-    digest
+    digest,
+    ps
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
 Suggests: 
