+++ R/vos_start.R
@@ -31,6 +31,7 @@ virtuoso_cache <- new.env()
 #' If no `virtuoso.ini` configuration file is provided, [vos_start()] will
 #' automatically attempt to configure one.  For more control over this,
 #' use [vos_configure()], see examples.
+#' @importFrom ps ps_pid
 #' @seealso [vos_install()]
 #' @examples \dontrun{
 #'
@@ -54,7 +55,9 @@ vos_start <- function(ini = NULL, wait = 30){
   p <- vos_process()
 
   if (inherits(p, "ps_handle")) {
-    message(paste("Returning existing virtuoso-t process with pid:", ps_pid(p)))
+    message(paste(
+      "Virtuoso is already running with pid:",
+      ps::ps_pid(p)))
     return(invisible(p))
   }
 
@@ -65,20 +68,20 @@ vos_start <- function(ini = NULL, wait = 30){
 
   ## Here we go time to start the process
   err <- file.path(vos_logdir(), "virtuoso.log")
-  p <- processx::process$new("virtuoso-t", c("-f", "-c", ini),
+  px <- processx::process$new("virtuoso-t", c("-f", "-c", ini),
                              stderr = err, stdout = "|",
                              cleanup = TRUE)
 
 
   ## Wait for status
-  message(p$format())
+  message(px$format())
   message("Server is now starting up, this may take a few seconds...\n")
   Sys.sleep(2)
-  vos_status(p, wait = wait)
+  vos_status(wait = wait)
 
 
   ##
-  invisible(p)
+  invisible(vos_process())
 }
 
 # FIXME check if virtuoso process is already running independently?
