+++ NAMESPACE
@@ -29,6 +29,7 @@ importFrom(processx,run)
 importFrom(ps,ps)
 importFrom(ps,ps_handle)
 importFrom(ps,ps_kill)
+importFrom(ps,ps_pid)
 importFrom(ps,ps_status)
 importFrom(rappdirs,app_dir)
 importFrom(utils,read.table)
