### 2019-02-04 5454265
merge all tests needing running vos into single location.
