### 2019-02-04 de3bdb6
xploring appveyor issues...
previous versions are passing on Windows locally so why is Appveyor unhappy?

Trying some alterations in the import script that may be a bit more robust.
