### 2019-02-04 30ffb2e
use ps functions directly to control virtuoso process (#24)
* migrate to ps-based service handling
* namespace fn calls directly from ps in tests
* merge all tests needing running vos into single location.
