+++ R/vos_log.R
@@ -20,8 +20,10 @@
 vos_log <- function(p = NA, collapse = NULL, just_errors = FALSE){
 
   p <- vos_process(p)
+  if(!inherits(p, "ps_handle")) return("")
+  err_file <- file.path(vos_logdir(), "virtuoso.log")
+  log <- readLines(err_file)
 
-  log <- readLines(p$get_error_file())
   if(just_errors){
     return(log[grepl("error", log)])
   }
