+++ R/vos_kill.R
@@ -10,18 +10,19 @@
 #' process id from the process table.
 #' @export
 #' @seealso [vos_start()]
-#' @aliases vos_kill vos_stop
+#' @aliases vos_kill
+#' @importFrom ps ps_kill
 #' @examples \dontrun{
 #' vos_start()
 #' vos_kill()
 #' }
 vos_kill <- function(p = NA){
   status <- vos_status(p)
-  if(status == "not detected")
-    stop("No active virtuoso process detected.")
+  if(is.null(status)){
+    message("No active virtuoso process detected.")
+    return(NULL)
+  }
   p <- vos_process(p)
-  p$kill()
-  rm("virtuoso_process", envir = virtuoso_cache)
+  ps::ps_kill(p)
 }
 
-vos_stop <- vos_kill
