+++ R/vos_process.R
@@ -5,6 +5,7 @@
 #' @inheritParams vos_kill
 #' @return returns the [processx::process()] object cached by [vos_start()]
 #' to control the external Virtuoso sever process from R.
+#' @importFrom ps ps_handle ps
 #' @export
 #' @examples \dontrun{
 #'
@@ -14,15 +15,26 @@
 #' }
 #'
 vos_process <- function(p = NA){
-  if (!inherits(p, "process")) {
-    p <- mget("virtuoso_process",
-              envir = virtuoso_cache,
-              ifnotfound = NA)[[1]]
+
+  ## p already is a handle to the process
+  if (inherits(p, "ps_handle")) {
+    return(p)
   }
-  if (!inherits(p, "process")) {
-    message(paste("No virtouso process found.",
-               "Try starting one with vos_start()\n"
-    ))
-  }
-  p
+
+  ## Otherwise, discover the pid
+  pid <- virtuoso_pid()
+
+  ## No pid means no running process
+  if (length(pid) == 0) return(NA)
+
+  ## Success. return a handle to this pid
+  ps::ps_handle(pid)
 }
+
+
+virtuoso_pid <- function(...){
+  x <- ps::ps(...)
+  x$pid[grepl("virtuoso-t", x$name)]
+}
+
+
