+++ tests/testthat/test-without-vos.R
@@ -1,9 +1,8 @@
 context("tests that do not need a server connection")
 
 test_that("vos_process errors when not cached", {
-
-  expect_message(vos_status())
-
-  expect_message(vos_process())
-#  "No virtuoso process found. Try starting one with vos_start()")
+  suppressMessages(vos_kill())
+  expect_message(vos_kill(), "No active virtuoso")
+  expect_null(vos_status())
+  expect_identical(vos_process(), NA)
 })

[…]
