### 2019-02-05 9903496
remaining issues from first reviewer
