+++ tests/testthat/setup-virtuoso.R
@@ -1,5 +1,6 @@
 
 message("Running one-time install of Virtuoso for tests...")
-vos_install(prompt = FALSE)
 
-Sys.sleep(20) # Give the installer a minute to finish...
+wait <- !has_virtuoso()
+vos_install(ask = FALSE)
+if(wait) Sys.sleep(20)
