+++ tests/testthat/test-utilities.R
@@ -14,8 +14,9 @@ test_that("we can guess extension", {
 })
 
 test_that("assert allowed dirs exits if server not connected ", {
-  skip_if(suppressWarnings(vos_status()))
+  skip_if(!is.null(suppressMessages(vos_status())))
 
-  expect_warning(assert_allowedDirs())
+  expect_warning(assert_allowedDirs(),
+                 "ensure working directory is in allowedDirs")
 
 })
