+++ tests/testthat/test-vos_install-errors.R
@@ -30,7 +30,7 @@ test_that("We get errors running windows installer on non-windows",
           {
             skip_on_cran()
             skip_on_os("windows")
-            expect_error(virtuoso:::vos_install_windows(prompt = FALSE))
+            expect_error(virtuoso:::vos_install_windows(ask = FALSE))
             expect_error(vos_uninstall_windows())
           })
 
@@ -40,8 +40,8 @@ test_that("We get errors running mac installer on non-mac",
   skip_on_cran()
   skip_on_os("mac")
   skip_on_os("windows")
-#   expect_error(virtuoso:::vos_install_osx(prompt = FALSE))
-#   expect_error(virtuoso:::vos_install_osx(brew = TRUE, prompt = FALSE))
+#   expect_error(virtuoso:::vos_install_osx(ask = FALSE))
+#   expect_error(virtuoso:::vos_install_osx(brew = TRUE, ask = FALSE))
 
   ## remarkably / scarily, this doesn't throw an error...
   #expect_error(virtuoso:::install_brew())
