+++ tests/testthat/setup-virtuoso.R
@@ -3,4 +3,4 @@ message("Running one-time install of Virtuoso for tests...")
 
 wait <- !has_virtuoso()
 vos_install(ask = FALSE)
-if(wait) Sys.sleep(20)
+if (wait) Sys.sleep(20)
