+++ R/vos_destroy_all.R
@@ -10,15 +10,16 @@
 #' @return [TRUE] if entirely successful in removing all files,
 #'  [FALSE] otherwise (invisibly).
 #' @export
-#' @examples \dontrun{
+#' @examples
+#' \dontrun{
 #' vos_destroy_all()
 #' }
-vos_destroy_all <- function(force = FALSE){
+vos_destroy_all <- function(force = FALSE) {
   s1 <- unlink(vos_db(), recursive = TRUE, force = force)
   s2 <- unlink(vos_cache(), recursive = TRUE, force = force)
   s3 <- unlink(vos_config(), recursive = TRUE, force = force)
   s4 <- unlink(vos_logdir(), recursive = TRUE, force = force)
-  invisible(sum(c(s1,s2,s3,s4)) == 0)
+  invisible(sum(c(s1, s2, s3, s4)) == 0)
 }
 
 #' Delete Virtuoso Database
@@ -28,10 +29,12 @@ vos_destroy_all <- function(force = FALSE){
 #' @param db_dir location of the directory to delete
 #' @export
 vos_delete_db <- function(ask = is_interactive(),
-                          db_dir = vos_db()){
-  continue = TRUE
-  if(ask)
+                          db_dir = vos_db()) {
+  continue <- TRUE
+  if (ask) {
     continue <- askYesNo("Are you sure?")
-  if(continue)
+  }
+  if (continue) {
     unlink(db_dir, recursive = TRUE)
+  }
 }
