+++ R/vos_query.R
@@ -11,14 +11,15 @@
 #' - <https://www.wikipedia.org/SPARQL>
 #' - <https://ropensci.github.io/rdflib/articles/rdf_intro.html>
 #'
-#' @examples \dontrun{
+#' @examples
+#' \dontrun{
 #' vos_start()
 #' con <- vos_connect()
-#'
+#' 
 #' # show first 4 triples in the database
 #' DBI::dbGetQuery(con, "SPARQL SELECT * WHERE { ?s ?p ?o } LIMIT 4")
 #' }
 #' @export
-vos_query <- function(con, query){
-  DBI::dbGetQuery(con, paste0("SPARQL ",  query))
+vos_query <- function(con, query) {
+  DBI::dbGetQuery(con, paste0("SPARQL ", query))
 }
