### 2019-02-05 c029636
minor tweaks
