+++ R/vos_install_linux.R
@@ -1,6 +1,6 @@
 
 vos_install_linux <- function() {
-  stop(paste(
+  message(paste(
     "Package does not support direct install of virtuoso",
     "from R on Linux systems. Please install virtuoso-opensource",
     "for your distribution. e.g. on Debian/Ubuntu systems, run",
