+++ R/vos_install_windows.R
@@ -1,5 +1,4 @@
 
-# FIXME support version number, add HTML munging to guess correct download URL from base URL?
 download_windows_installer <- function(version = "7.2.5") {
   exe <- "Virtuoso_OpenSource_Server_7.20.x64.exe"
   download_url <- paste0(
