### 2019-02-06 #271 @cboettig
@edgararuiz Okay, can you install `install_github("cboettig/virtuoso@ropensci-review")` and give it a second run?

@czeildi Thanks for the stellar review, I think I've addressed each of the issues you've raised now as well.  you can see all the changes I've made in response to both of your reviews so far in this open PR, https://github.com/cboettig/virtuoso/pull/25, hopefully makes it somewhat easier to track.  For Ildikó 's review, I've summarized the changes point-by-point in the associated issue.

Thanks both for all the super constructive testing. Edgar, hope we've got this working more robustly on Windows now as well, but looking forward to your testing as this kind of thing is particularly tricky to get the cross-platform details working with so much system-specific operations, (even once I've satisfied Appveyor and my local Windows 10 box)
