### 2019-02-09 #271 @czeildi
@cboettig  I checked your pull request, looks great, now I am happy to recommend approving virtuoso.

Thanks for replying point-by-point and in detail.

I found one small typo: in `vos_connect.R` Default parameters are *approporiate*

Regarding the datalake article:
- the readme works for me fully
- the problem is not related to that chunk: my virtuoso process is killed after some time seemingly without reason (Running only `vos_status()` repeatedly and nothing else it reported the process is not running after some time). I think we can let this go at this time.
