### 2019-02-17 #271 @edgararuiz
@cboettig -

I'm getting the follow when using the new branch:

```
vos_install()
downloading Virtuoso_OpenSource_Server_7.20.x64.exe ...
 Error in curl::curl_fetch_disk(url, dest) :
  Timeout was reached: Connection timed out after 10625 milliseconds
