### 2019-02-17 #271 @cboettig
Thanks, hm, can you double check you are using the version installed from
the branch and not from master?

On Sun, Feb 17, 2019 at 11:12 AM Edgar Ruiz <notifications@github.com>
wrote:

> @cboettig <https://github.com/cboettig> -
>
> I'm getting the follow when using the new branch:
>
> vos_install()
> downloading Virtuoso_OpenSource_Server_7.20.x64.exe ...
>  Error in curl::curl_fetch_disk(url, dest) :
>   Timeout was reached: Connection timed out after 10625 milliseconds
>
