### 2019-02-18 #271 @cboettig
@edgararuiz very weird.  Still works for me, and the current branch also has a fallback mechanism it should use if the binary really was not reproducible, so this error looks like only what I can get on master branch.  You also restarted R as well?

Can you show me what you see by running `virtuoso:::download_windows_installer` (no parens, to show function def).  The new branch should show like this:

```
virtuoso:::download_windows_installer
function(version = "7.2.5"){
  exe <- "Virtuoso_OpenSource_Server_7.20.x64.exe"
  download_url <- paste0("https://sourceforge.net/projects/virtuoso/",
                         "files/virtuoso/7.2.5/",
                         "2018_08_28_",
                         "Virtuoso_OpenSource_Server_7.2.x64.exe/download")
  fallback_url <- paste0("https://github.com/cboettig/virtuoso/releases/",
                             "download/v0.1.1/Virtuoso_OpenSource_Server_7.20.x64.exe")
  installer <- normalizePath(file.path(
    tempdir(),
    exe),
    mustWork = FALSE)
  message(paste("downloading", exe,  "..."))
  download_fallback(download_url, installer, fallback_url)
  installer
}
```

you might also check if you can access that download URL?  Is it possible you have some firewall settings that prevent downloads from sourceforge (and github??)

(note that on the other hand, the sourceforge URL in the branch on master has moved, and had not fallback URL, so it should create the exact timeout error you see).  Sorry for the trouble!
