### 2019-02-19 03c4cf4
Edits addressing rOpenSci peer review (#25)
* edits for rOpenSci review (closes #26)
* use ps functions directly to control virtuoso process (#24) (closes #23)
* enable a fallback url for downloading
* run styler::style_pkg()
* update vignette
