### 2019-02-20 #271 @stefaniebutland
@cboettig You're always welcome to write a post, but I recall you said earlier that you prefer to save your writing energy for papers. 🙂
