### 2019-02-20 44f6b0b
update README, DESCRIPTION links, and codemeta
