### 2019-02-20 dbe8b90
rebuild docs for ropensci-side
