### 2019-02-22 #271 @cboettig
Thanks @melvidoni !  I've transferred the repo and updated links.  Preparing CRAN submission...
