### 2019-02-22 f6afa52
:bug: fix: avoid gc of virtuoso processx
We need to cache processx process so server persists, even though we can rediscover it by ps.  Otherwise, garbage collection seems to think we're done with the process and kills it.

Also, update vignette now that this is all working.
