+++ README.md
@@ -1,23 +1,14 @@
 
-# virtuoso <img src="man/figures/logo.svg" align="right" alt="" width="120" />
+virtuoso <img src="man/figures/logo.svg" align="right" alt="" width="120" />
+============================================================================
 
-[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![Travis build
-status](https://travis-ci.org/ropensci/virtuoso.svg?branch=master)](https://travis-ci.org/ropensci/virtuoso)
-[![AppVeyor build
-status](https://ci.appveyor.com/api/projects/status/github/cboettig/virtuoso?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/virtuoso)
-[![Coverage
-status](https://codecov.io/gh/ropensci/virtuoso/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/virtuoso?branch=master)
-[![CRAN
-status](https://www.r-pkg.org/badges/version/virtuoso)](https://cran.r-project.org/package=virtuoso)
-[![](http://badges.ropensci.org/271_status.svg)](https://github.com/ropensci/software-review/issues/271)
+[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing) [![Travis build status](https://travis-ci.org/ropensci/virtuoso.svg?branch=master)](https://travis-ci.org/ropensci/virtuoso) [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/virtuoso?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/virtuoso) [![Coverage status](https://codecov.io/gh/ropensci/virtuoso/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/virtuoso?branch=master) [![CRAN status](https://www.r-pkg.org/badges/version/virtuoso)](https://cran.r-project.org/package=virtuoso) [![](http://badges.ropensci.org/271_status.svg)](https://github.com/ropensci/software-review/issues/271)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+The goal of virtuoso is to provide an easy interface to Virtuoso RDF database from R.
 
-The goal of virtuoso is to provide an easy interface to Virtuoso RDF
-database from R.
-
-## Installation
+Installation
+------------
 
 You can install the development version of virtuoso from GitHub with:
 
@@ -25,16 +16,14 @@ You can install the development version of virtuoso from GitHub with:
 remotes::install_github("ropensci/virtuoso")
 ```
 
-## Getting Started
+Getting Started
+---------------
 
 ``` r
 library(virtuoso)
 ```
 
-For Mac users, `virtuoso` package includes a utility function to install
-and configure a local Virtuoso Open Source instance using Homebrew.
-Otherwise, simply install the Virtuoso Open Source edition for your
-operating system.
+For Mac users, `virtuoso` package includes a utility function to install and configure a local Virtuoso Open Source instance using Homebrew. Otherwise, simply install the Virtuoso Open Source edition for your operating system.
 
 ``` r
 vos_install()
@@ -45,9 +34,9 @@ We can now start our Virtuoso server from R:
 
 ``` r
 vos_start()
-#> PROCESS 'virtuoso-t', running, pid 3209.
+#> PROCESS 'virtuoso-t', running, pid 1126.
 #> Server is now starting up, this may take a few seconds...
[…]
