+++ vignettes/virtuoso.Rmd
@@ -12,7 +12,8 @@ vignette: >
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  eval = !virtuoso:::is_windows()
 )
 ```
 
