### 2019-02-25 f3715d2
do not attempt to install virtuoso server on CRAN
