+++ tests/testthat/setup-virtuoso.R
@@ -1,6 +1,8 @@
 
 message("Running one-time install of Virtuoso for tests...")
-
 wait <- !has_virtuoso()
-vos_install(ask = FALSE)
+
+if (identical(Sys.getenv("NOT_CRAN"), "true"))
+  vos_install(ask = FALSE)
 if (wait) Sys.sleep(20)
+
