### 2019-02-25 b9fbeda
hmm, is appveyor skipping skip_on_cran?
