+++ appveyor.yml
@@ -15,7 +15,7 @@ install:
   - ps: Bootstrap
 
 cache:
-#  - C:\RLibrary
+  - C:\RLibrary
 
 # Adapt as necessary starting from here
 
