### 2019-02-25 c5c876c
in which we manually tell Appveyor it's not CRAN
Thanks to the brilliant @ldecicco-USGS for this solution! :sparkles:
