+++ tests/testthat/setup-virtuoso.R
@@ -4,7 +4,7 @@ testthat::setup({
   message("Running one-time install of Virtuoso for tests...")
   wait <- !has_virtuoso()
 
-  #testthat::skip_on_cran()
+  testthat::skip_on_cran()
   vos_install(ask = FALSE)
   if (wait) Sys.sleep(20)
 
