+++ tests/testthat/setup-virtuoso.R
@@ -1,8 +1,12 @@
 
+
+testthat::setup({
 message("Running one-time install of Virtuoso for tests...")
 wait <- !has_virtuoso()
 
-if (identical(Sys.getenv("NOT_CRAN"), "true"))  vos_install(ask = FALSE)
+# testthat::skip_on_cran()
+# if (identical(Sys.getenv("NOT_CRAN"), "true"))
+  vos_install(ask = FALSE)
 if (wait) Sys.sleep(20)
 
-
+})
