+++ appveyor.yml
@@ -2,7 +2,7 @@
 environment:
   global:
     USE_RTOOLS: true
-    NOT_CRAN: true
+    NOT_CRAN: "true"
 
 # Download script file from GitHub
 init:
