+++ appveyor.yml
@@ -15,8 +15,8 @@ init:
 install:
   - ps: Bootstrap
 
-cache:
-  - C:\RLibrary
+#cache:
+#  - C:\RLibrary
 
 # Adapt as necessary starting from here
 
