+++ appveyor.yml
@@ -1,6 +1,7 @@
 # DO NOT CHANGE the "init" and "install" sections below
 environment:
   global:
+    NOT_CRAN: "true"
     USE_RTOOLS: true
 
 # Download script file from GitHub
