### 2019-02-25 66b281b
vos_query tests failing on appveyor
