### 2019-02-26 c77d8ca
skip_on_cran does not work in setup blocks
