+++ tests/testthat/test-vos_query.R
@@ -2,9 +2,12 @@ context("test vos_query")
 
 testthat::setup({
 
-  skip_on_cran()
-  vos_start(wait = 120)
-  })
+  ## skip_on_cran() not applicable to setup blocks!
+  if (identical(Sys.getenv("NOT_CRAN"), "true")){
+    vos_start(wait = 120)
+  }
+
+})
 
 
 
