+++ README.Rmd
@@ -10,7 +10,7 @@ output: github_document
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/virtuoso?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/virtuoso)
 [![Coverage status](https://codecov.io/gh/ropensci/virtuoso/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/virtuoso?branch=master)
 [![CRAN status](https://www.r-pkg.org/badges/version/virtuoso)](https://cran.r-project.org/package=virtuoso) 
-[![](http://badges.ropensci.org/271_status.svg)](https://github.com/ropensci/software-review/issues/271)
+[![Peer review](http://badges.ropensci.org/271_status.svg)](https://github.com/ropensci/software-review/issues/271)
 
  
 <!-- README.md is generated from README.Rmd. Please edit that file -->
