### 2019-02-26 be96043
Patch Appveyor and CRAN winbuilder testing (#27)
* appveyor :fire:
* in which we manually tell Appveyor it's not CRAN
