+++ appveyor.yml
@@ -2,6 +2,7 @@
 environment:
   global:
     USE_RTOOLS: true
+    NOT_CRAN: "true"
 
 # Download script file from GitHub
 init:
@@ -14,8 +15,8 @@ init:
 install:
   - ps: Bootstrap
 
-cache:
-  - C:\RLibrary
+#cache:
+#  - C:\RLibrary
 
 # Adapt as necessary starting from here
 
