### 2019-02-26 94baff8
add NEWS.md
