+++ NEWS.md
@@ -1,0 +1,3 @@
+# virtuoso 0.1.2
+
+* Added a `NEWS.md` file to track changes to the package.
